sxs-chat-integration/requirements.txt
flask==2.0.1
gunicorn==20.1.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.31.0
google-cloud-dialogflow-cx==1.12.0